requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
filelock==3.13.1
fsspec==2023.12.2
huggingface-hub==0.19.4
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
shapely==2.0.2
sympy==1.12
tokenizers==0.15.0
torch==2.1.1
tqdm==4.66.1
transformers==4.36.0
typing-extensions==4.9.0
urllib3==2.1.0
